+++ adminer/foreign.inc.php
@@ -39,10 +39,10 @@ page_header(lang('Foreign key'), $error, array("table" => $TABLE), h($TABLE));
 
 if ($_POST) {
 	ksort($row["source"]);
-	if ($_POST["add"]) {
-		$row["source"][] = "";
-	} elseif ($_POST["change"] || $_POST["change-js"]) {
+	if ($_POST["change"] || $_POST["change-js"]) {
 		$row["target"] = array();
+	} else {
+		$row["source"][] = "";
 	}
 } elseif ($name != "") {
 	$foreign_keys = foreign_keys($TABLE);