+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Adminer dev
 - Select: Disable Ctrl+click inline edit without UPDATE privilege
 - Export: Remember unchecked objects (regression from 5.0.6)
+- Foreign key: Display new field in case of an error
 - PostgreSQL: Order NULL last
 - PostgreSQL: Display all SQL command warnings and only once